1.表格显示顺序 2.表格页面 图片大小作显示限制 3.订单index，需显示样品小图 4.原材料仓库单，保存后 小数变成了 整数 BUG 5.对话框打开后，鼠标焦点应直接 定位在第一个可编辑的input上 6.提示对话框打开后，焦点应定位在提示对话框上

+++ src/com/fuwei/entity/ordergrid/StoreOrderDetail.java
@@ -3,7 +3,7 @@ package com.fuwei.entity.ordergrid;
 public class StoreOrderDetail {
 	private String color;//色号
 	private Integer material;//材料
-	private int quantity;//总数量
+	private double quantity;//总数量
 	private String yarn;//标准样纱
 	private Integer factoryId;//领取人 ，必填 （工厂）
 	
@@ -26,10 +26,11 @@ public class StoreOrderDetail {
 	public void setMaterial(Integer material) {
 		this.material = material;
 	}
-	public int getQuantity() {
+	
+	public double getQuantity() {
 		return quantity;
 	}
-	public void setQuantity(int quantity) {
+	public void setQuantity(double quantity) {
 		this.quantity = quantity;
 	}
 	public String getYarn() {
